changed styleExports.js to .ts and added a type for style objects and removed unused ts directive
